refactor(frontend): split Navbar, SSR into layout.server.ts & page.server.ts
